pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sentineldb-env"
version = "0.1.0"
description = "SentinelDB OpenEnv environment for multi-agent cyber defense."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "madhu" }
]
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn>=0.27.0",
  "pydantic>=2.6.0",
  "httpx>=0.27.0",
  "pytest>=7.4.0",
  "matplotlib>=3.8.0",
  "pandas>=2.2.0",
  "openenv-core==0.2.3",
  "trl>=0.27.1",
  "transformers>=4.40.0",
  "datasets>=2.18.0",
  "torch>=2.2.0",
]

[project.scripts]
sentineldb-server = "server.app:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["sentineldb_env*", "server*"]
